added user access control - ProductColor

+++ src/main/java/pl/coderslab/controller/ProductColorController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import pl.coderslab.model.ProductColor;
 import pl.coderslab.service.ProductColorService;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 
@@ -31,7 +32,11 @@ public class ProductColorController {
 
 
     @GetMapping("/add")
-    public String add(Model model) {
+    public String add(Model model, HttpSession session) {
+        Long userGroupId = (Long) session.getAttribute("userGroup");
+        if (userGroupId == 2) {
+            return "redirect:/noAccess";
+        }
         model.addAttribute("productColor", new ProductColor());
         return "productColors/add";
     }
@@ -48,7 +53,11 @@ public class ProductColorController {
 
 
     @GetMapping("/edit/{id}")
-    public String update(@PathVariable Long id, Model model) {
+    public String update(@PathVariable Long id, Model model, HttpSession session) {
+        Long userGroupId = (Long) session.getAttribute("userGroup");
+        if (userGroupId == 2) {
+            return "redirect:/noAccess";
+        }
         model.addAttribute("productColor", productColorService.findById(id));
         return "productColors/edit";
     }
@@ -65,7 +74,11 @@ public class ProductColorController {
 
 
     @RequestMapping("/delete/{id}")
-    public String delete(@PathVariable Long id) {
+    public String delete(@PathVariable Long id, HttpSession session) {
+        Long userGroupId = (Long) session.getAttribute("userGroup");
+        if (userGroupId == 2) {
+            return "redirect:/noAccess";
+        }
         try {
             productColorService.delete(id);
             return "redirect:/productColors/all?deleted=true";
